Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,7 +1,7 @@
 #include<stdio.h>
 int main(){
-    int a , b , c , sum , average;
-    scanf("%d %d %d" , &a , &b , &c);
+    double a , b , c , sum , average;
+    scanf("%lf %lf %lf" , &a , &b , &c);
     sum = a + b + c;
     average = sum / 3;
     printf("Average: %.2lf " , average);
